webapp/backend/requirements.txt
flask==3.0.3
weaviate-client==4.6.5
boto3==1.34.144
flask-cors==4.0.1
